add a custom Maven Jenkins slave Docker image
We need to run tests on various OpenShift versions. For each of them,
we need to use the corresponding version of the `oc` binary. Upstream
is working on a better solution for this (versioned Jenkins slave
images), but until that is available (and pervasive!), this image
is a good solution.

It works like this: the entrypoint of the Docker image is changed
to a custom shell script. This script finds out the current OpenShift
version (by querying the API), downloads the correct `oc` binary
from GitHub, prepends it to `PATH` and `exec`s the original script
(which starts the Jenkins slave process).

There are two downsides:

1. The script has a hardcoded list of known OpenShift versions
   and corresponding `oc` archive URLs. If there's a new OpenShift
   release, the script needs to be updated and the image rebuilt.
2. We don't download the exactly precisely same version of the `oc`
   binary, just the same major.minor. This should be good enough
   for all practical purposes.